week1/Job-skill-scraper/pyproject.toml
[project]
name = "job-skill-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-core>=1.2.23",
    "langchain-ollama>=1.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.0",
    "urllib3>=2.6.3",
]
